Added move image to different album

+++ src/components/deleteItems.js
@@ -3,16 +3,25 @@ import React from 'react';
 import { photoBoothFirestore } from './../firebase/config';
 
 // function to delete an image
-const DeleteItem = ( field, user, selectedItem, setSelectedItem ) => {
+const DeleteItem = ( method, albumName, field, userId, selectedItem, setSelectedItem ) => {
     // references
-    const collectionRef = photoBoothFirestore.collection(`${user}`); // create reference to user's collection in firestore
+    const collectionRef = photoBoothFirestore.collection(`${userId}`); // create reference to user's collection in firestore
     const documents = collectionRef.where(field,'==', selectedItem); // set the query field to the field of the selected file
 
     // get the documents with the set query
     documents.get().then( (querySnapshot) => {
       // for each document found
       querySnapshot.forEach( (doc) => {
-        doc.ref.delete(); // delete the document
+
+        // If the deletion was of an image
+        if(method == 'delete'){
+          doc.ref.delete(); // delete the document
+        }
+        // otherwise if the method is update
+        else if (method == 'update') {
+          doc.ref.update({ album: albumName }); // update the document
+        }
+
       });
     }).then(() => {
       // If the deletion was of an image
